added namespace to avoid conflicts with Official gem
