+++ lib/airops.rb
@@ -1,0 +1,4 @@
+require "airops/openai"
+
+module Airops
+end
