+++ lib/airops/openai/compatibility.rb
@@ -1,0 +1,11 @@
+module Airops
+  module Ruby
+    module OpenAI
+      VERSION = ::Airops::OpenAI::VERSION
+
+      Error = ::Airops::OpenAI::Error
+      ConfigurationError = ::Airops::OpenAI::ConfigurationError
+      Configuration = ::Airops::OpenAI::Configuration
+    end
+  end
+end
