+++ lib/airops/openai/version.rb
@@ -1,0 +1,5 @@
+module Airops
+  module OpenAI
+    VERSION = "3.7.0".freeze
+  end
+end
