+++ lib/airops/ruby/openai.rb
