--- lib/openai/compatibility.rb
@@ -1,9 +1,0 @@
-module Ruby
-  module OpenAI
-    VERSION = ::OpenAI::VERSION
-
-    Error = ::OpenAI::Error
-    ConfigurationError = ::OpenAI::ConfigurationError
-    Configuration = ::OpenAI::Configuration
-  end
-end
