--- lib/openai/version.rb
@@ -1,3 +1,0 @@
-module OpenAI
-  VERSION = "3.7.0".freeze
-end
