+++ spec/airops/openai/client/embeddings_spec.rb
@@ -1,9 +1,9 @@
-RSpec.describe OpenAI::Client do
+RSpec.describe Airops::OpenAI::Client do
   describe "#embeddings", :vcr do
     let(:input) { "The food was delicious and the waiter..." }
     let(:cassette) { "#{model} embeddings #{input}".downcase }
     let(:response) do
-      OpenAI::Client.new.embeddings(
+      Airops::OpenAI::Client.new.embeddings(
         parameters: {
           model: model,
           input: input
