+++ spec/airops/openai_spec.rb
@@ -1,0 +1,54 @@
+RSpec.describe Airops::OpenAI do
+  it "has a version number" do
+    expect(Airops::OpenAI::VERSION).not_to be nil
+  end
+
+  describe "#configure" do
+    let(:access_token) { "abc123" }
+    let(:api_version) { "v2" }
+    let(:organization_id) { "def456" }
+    let(:custom_uri_base) { "ghi789" }
+    let(:custom_request_timeout) { 25 }
+
+    before do
+      Airops::OpenAI.configure do |config|
+        config.access_token = access_token
+        config.api_version = api_version
+        config.organization_id = organization_id
+      end
+    end
+
+    it "returns the config" do
+      expect(Airops::OpenAI.configuration.access_token).to eq(access_token)
+      expect(Airops::OpenAI.configuration.api_version).to eq(api_version)
+      expect(Airops::OpenAI.configuration.organization_id).to eq(organization_id)
+      expect(Airops::OpenAI.configuration.uri_base).to eq("https://api.openai.com/")
+      expect(Airops::OpenAI.configuration.request_timeout).to eq(0)
+    end
+
+    context "without an access token" do
+      let(:access_token) { nil }
+
+      it "raises an error" do
+        expect { Airops::OpenAI::Client.new.completions }.to raise_error(Airops::OpenAI::ConfigurationError)
+      end
+    end
+
+    context "with custom timeout and uri base" do
+      before do
+        Airops::OpenAI.configure do |config|
+          config.uri_base = custom_uri_base
+          config.request_timeout = custom_request_timeout
+        end
+      end
+
+      it "returns the config" do
+        expect(Airops::OpenAI.configuration.access_token).to eq(access_token)
+        expect(Airops::OpenAI.configuration.api_version).to eq(api_version)
+        expect(Airops::OpenAI.configuration.organization_id).to eq(organization_id)
+        expect(Airops::OpenAI.configuration.uri_base).to eq(custom_uri_base)
+        expect(Airops::OpenAI.configuration.request_timeout).to eq(custom_request_timeout)
+      end
+    end
+  end
+end
